+++ dave/common/server_type.py
@@ -1,9 +1,10 @@
 from enum import Enum
 from typing import Union
 
-SERVER_TYPE = None #type: Union[ServerType, None]
+SERVER_TYPE = None  # type: Union[ServerType, None]
+
 
 class ServerType(Enum):
     GDB = "gdb"
     LLDB = "lldb"
-    PDB = "pdb"
+    PYTHON = "pdb"