+++ examples/custom_containers/__init__.py
@@ -1,0 +1,6 @@
+from custom_containers.example import (
+    DaveCustomContainerPtr,
+    DaveCustomContainerPtrPtr,
+    DaveCustomContainerVecRef,
+    DaveCustomInterleavedContainerVec,
+)