+++ examples/np.py
@@ -1,7 +1,7 @@
 import numpy as np
 import cmath
 
-import dave.server.debuggers.pdb as pydave
+# import dave.server.debuggers.pdb as pydave
 
 
 BLOCK_SIZE = 4096